re-organize utils subpack
