+++ utils/zelux.py
@@ -1,7 +1,82 @@
 import numpy as np
-from scipy import ndimage
+
 from thorlabs_tsi_sdk.tl_camera import TLCameraSDK
 
+from .misc import get_params
+
+class ZeluxCamera():
+    
+    def __init__(self, **config):
+        print("Zelux camera initiliazed - use init_cam() to arm and trigger it and get() to get frames")
+        get_params(self, **config)
+    
+    def set_roi(self):
+        """ Calculates the Region of interest. The user gives a window size and x, y offsets from
+            the sensor center
+
+        Returns
+        -------
+        roi (tuple or int)
+        """
+        if type(self.roi_size) is tuple:
+            width, height = self.roi_size
+        else:
+            width = self.roi_size
+            height = width
+            
+        offset_x, offset_y = self.offset
+        middle_x = int(1440 / 2) + offset_x
+        middle_y = int(1080 / 2) - offset_y
+
+        roi = (middle_x - int(width/ 2), 
+            middle_y - int(height / 2), 
+            middle_x + int(width / 2), 
+            middle_y + int(height / 2))
+        
+        return roi
+    
+    def init_cam(self):
+        """ Initializes and sets camera parameters
+        """
+        # camera instance
+        self.sdk = TLCameraSDK()
+        available_cameras = self.sdk.discover_available_cameras()
+        self.camera = self.sdk.open_camera(available_cameras[0])
+
+        # configure
+        self.camera.exposure_time_us = self.exposure_time  # set exposure to 11 ms
+        self.camera.frames_per_trigger_zero_for_unlimited = 0  # start camera in continuous mode
+        self.camera.image_poll_timeout_ms = self.timeout  # 1 second polling timeout
+        self.camera.roi = self.set_roi()
+
+        # set binning for camera macropixels
+        (self.camera.binx, self.camera.biny) = (self.bins, self.bins)
+
+        if self.camera.gain_range.max > self.gain:
+            db_gain = self.gain
+            gain_index = self.camera.convert_decibels_to_gain(db_gain)
+            self.camera.gain = gain_index
+
+        # arm - trigger
+        self.camera.arm(2)
+        self.camera.issue_software_trigger()
+        
+        return self.camera
+    
+    def close_cam(self):
+        self.camera.disarm()
+        self.camera.dispose()
+        self.sdk.dispose()  
+        
+    def get(self):
+        """ Get frame from zelux thorlabs camera
+        """
+        frame = self.camera.get_pending_frame_or_null()
+        image_buffer = np.copy(frame.image_buffer)
+        
+        return image_buffer
+    
+    
 
 def set_roi(roi_size, roi_off):
     offset_x, offset_y = roi_off
@@ -86,84 +161,4 @@ class Cam:
 
                 camera.disarm()
             return frame
-
-# def get_frame_binned(roi, bins, num_of_frames=1, exposure_time=5000, gain=1, timeout=1000, return_roi=False):
-
-#     frame = get_interferogram(roi=roi,
-#                               bins=(bins, bins),
-#                               num_of_frames=num_of_frames,
-#                               exposure_time=exposure_time,
-#                               gain=gain,
-#                               timeout=timeout, 
-#                               return_roi=return_roi)
-
-#     return frame
-
-# class InitCamera():
-#     def __init__(self, roi_size=100, roi_off=(0, 0), bins=1, exposure_time=50, gain=1, timeout=100):    
-#         # camera settings
-#         self.roi_size = roi_size
-#         self.offset_x, self.offset_y = roi_off # roi offset from center
-#         self.bins = bins
-#         self.exposure_time = exposure_time
-#         self.gain = gain
-#         self.timeout = timeout
-    
-#     def set_roi(self):
-#         """ Calculates the Region of interest. The user gives a window size and x, y offsets from
-#             the sensor center
-
-#         Returns
-#         -------
-#         roi (tuple)
-#         """
-
-#         middle_x = int(1440 / 2) + self.offset_x
-#         middle_y = int(1080 / 2) - self.offset_y
-
-#         roi = (middle_x - int(self.roi_size / 2), 
-#                middle_y - int(self.roi_size / 2), 
-#                middle_x + int(self.roi_size / 2), 
-#                middle_y + int(self.roi_size / 2))
         
-#         return roi
-        
-#     def config(self):
-#         """_summary_
-
-#         Returns
-#         -------
-#         _type_
-#             _description_
-#         """
-#         # camera instance
-#         self.sdk = TLCameraSDK()
-#         available_cameras = self.sdk.discover_available_cameras()
-#         self.camera = self.sdk.open_camera(available_cameras[0])
-
-#         # configure
-#         self.camera.exposure_time_us = self.exposure_time  # set exposure to 11 ms
-#         self.camera.frames_per_trigger_zero_for_unlimited = 0  # start camera in continuous mode
-#         self.camera.image_poll_timeout_ms = self.timeout  # 1 second polling timeout
-#         self.camera.roi = self.set_roi()
-
-#         # set binning for camera macropixels
-#         (self.camera.binx, self.camera.biny) = (self.bins, self.bins)
-
-#         if self.camera.gain_range.max > self.gain:
-#             db_gain = self.gain
-#             gain_index = self.camera.convert_decibels_to_gain(db_gain)
-#             self.camera.gain = gain_index
-
-#         # arm - trigger
-#         self.camera.arm(2)
-#         self.camera.issue_software_trigger()
-        
-#         return self.camera
-    
-#     def destroy(self):
-#         self.camera.disarm()
-#         self.camera.dispose()
-#         self.sdk.dispose()
-        
-
